+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix action disabling with mixins [#298](https://github.com/moleculerjs/moleculer/issues/298)
 - Fix metrics options and add findNextActionEndpoint to index.d.ts
 - update dependencies
+- set `maxReconnectAttempts` to `-1` in NATS client to try reconnecting continuously
 
 --------------------------------------------------
 <a name="0.12.6"></a>